+++ R/policy.R
@@ -41,7 +41,9 @@ policy <- function(x) x$solution$policy
 
 .policy_MDP_from_POMDP <- function(x) {
   pg <- x$solution$pg
-  bs <- x$observation_prob[['*']]
+  
+  ## all observation_probs should be the same!
+  bs <- x$observation_prob[[1L]]
   
   # create a list ith epochs
   lapply(